Build sitemaps (#51)
* add build sitemaps script

* regenerated docs

* Add files via upload

* Add files via upload

* Delete google319a4ae362a3a60e.html

* Add files via upload

* avoid scss variable if it doesnt exist (#54)

* can now pass path for automatic docs route generation (#57)

* This diff support creation of tables of content for routes described in jsRoutes. (#56)

Previously, jsRoutes was taken as is - there was no reprocessing at all. All routes that needed to appear in the table of contents had to be in mdRoutes. The issue there is that mdRoutes mixed 2 kind of routes - those with documentations (as intended) and those with standalone components.
With this PR, we can go back to a model where mdRoutes contains only routes with markdown files (and can easily be generated instead of manually created)

* add build sitemaps script

* regenerated docs

* update sitemap syntax

* before rebasing