Add entry in babel for default env value "development" as alias to "es5" (#161)
The BABEL_ENV environment variable is used to control the babel configuration from ocular-dev-tools.  It is standard to have the values `production`, `development`, `test`.  In ocular this environment value controls the target format with values `es5`, `es6`, `esm`.

In order to support the standard `development` value used when developing locally this adds an entry for `development` which aliases `es5`.  This is useful in conjunction with `@babel/register`.

This simply supports the default value expected in babel workflows.  This behavior can always be overwritten by setting `BABEL_ENV` explicitly.

https://new.babeljs.io/docs/en/next/babelrc.html
https://babeljs.io/docs/en/options